moved location based functions to their own file.  Added unit test.
