+++ ts/packages/commonUtils/src/indexNode.ts
@@ -38,3 +38,5 @@ export {
 } from "./mimeTypes.js";
 
 export { getObjectProperty, setObjectProperty } from "./objectProperty.js";
+
+export * from "./location.js";
